implement complete user registration flow

+++ src/utils/cloudinary.js
@@ -2,8 +2,8 @@
 // cloudinary is a cloud-based media storage service used to upload, store and manage images, videos and other files, instead of storing uploaded files on our own server, we upload them to cloudinary, cloudinary safely stores the files and returns a url
 // we then save that url in our mongodb database instead of saving the actual file, making our application faster and reducing server storage
 
-import { v2 as cloudinary } from 'cloudinary';  // import the cloudinary sdk, used to upload images and videos to cloudinary cloud storage
-import fs from 'fs';    // node's built-in file system module, used to delete temporary files stored on our local server
+import { v2 as cloudinary } from 'cloudinary'; // import the cloudinary sdk, used to upload images and videos to cloudinary cloud storage
+import fs from 'fs'; // node's built-in file system module, used to delete temporary files stored on our local server
 
 // configure cloudinary using credentials stored in environment variables, this connects our backend application to our cloudinary account
 cloudinary.config({
@@ -16,19 +16,23 @@ cloudinary.config({
 // localFilePath -> path of the file stored temporarily on our server
 const uploadOnCloudinary = async (localFilePath) => {
     try {
-        if (!localFilePath) return null;    // if no file path is provided, nothing can be uploaded
-        
+        if (!localFilePath) return null; // if no file path is provided, nothing can be uploaded
+
         // upload the file to cloudinary
         // resource_type: 'auto' automatically detects whether the file is an image, video or another supported file type
         const response = await cloudinary.uploader.upload(localFilePath, {
             resource_type: 'auto',
         });
-        console.log('File is upload on cloudinary', response.url);
-        return response;    // return the complete response object received from cloudinary
+        // the file has now been successfully uploaded to cloudinary, it it no longer needed on our local server, so delete the temporary file
+        // this prevents unnecessary files from accumulating inside the public/temp folder
+        fs.unlinkSync(localFilePath);
+        return response; // return the complete response object received from cloudinary
     } catch (error) {
-        fs.unlinkSync(localFilePath);   // if upload fails delete the temporary file, so unnecessary files don't remain on our server
+        if (localFilePath) {
+            fs.unlinkSync(localFilePath); // if an error occurs during the upload, delete the temporary local file to avoid leaving unused files on the server
+        }
         return null;
     }
 };
 
-export { uploadOnCloudinary }
+export { uploadOnCloudinary };
